requirements.txt
absl-py==1.4.0
astunparse==1.6.3
attrs==23.1.0
autopep8==2.0.2
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
contourpy==1.0.7
cycler==0.11.0
dask==2023.4.1
demes==0.2.2
docopt==0.6.2
flatbuffers==23.5.26
fonttools==4.39.3
fsspec==2023.4.0
gast==0.3.3
google-auth==2.17.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
humanize==4.6.0
idna==3.4
imageio==2.28.1
importlib-metadata==6.6.0
importlib-resources==5.12.0
jax==0.4.8
joblib==1.2.0
jsonschema==4.17.3
kastore==0.3.2
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
lazy-loader==0.2
libclang==16.0.0
llvmlite==0.38.1
locket==1.0.0
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.4.2
ml-dtypes==0.1.0
msprime==1.2.0
networkx==3.1
newick==1.9.0
numba==0.55.2
numpy==1.22.0
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.1
pandas==1.2.4
partd==1.4.0
Pillow==9.5.0
pkgutil-resolve-name==1.3.10
protobuf==3.20.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.10.0
pyparsing==3.0.9
pyrsistent==0.19.3
pyslim==0.700
python-dateutil==2.8.2
python-version==0.0.2
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
requests==2.30.0
requests-oauthlib==1.3.1
rsa==4.9
ruamel.yaml==0.17.22
ruamel.yaml.clib==0.2.7
scikit-allel==1.3.5
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.9.1
seaborn==0.12.2
shap==0.41.0
six==1.15.0
slicer==0.0.7
svgwrite==1.4.3
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.1.0
tifffile==2023.4.12
tomli==2.0.1
toolz==0.12.0
tqdm==4.65.0
tskit==0.4.1
typing-extensions==3.7.4.3
tzdata==2023.3
urllib3==2.0.2
Werkzeug==2.3.3
wrapt==1.12.1
yarg==0.1.9
zipp==3.15.0
